+++ R/matchFormula.R
@@ -1,25 +1,34 @@
-#' @title formula matching
+#' @title Chemical Formula Matching
 #'
 #' @name matchFormula
 #'
 #' @description
 #'
-#' The `matchFormula` method matches chemical formula from different inputs
+#' The `matchFormula` method matches chemical formulas from different inputs
 #' (parameter `query` and `target`). Before comparison all formulas are
-#' normalized using [MetaboCoreUtils::standardizeFormula()]. Inputs can be either
-#' a `character` or `data.frame` containing a column with formulas. In case of
-#' a `data.frame` as input the parameter `formulaColName` needs to specified.
+#' normalized using [MetaboCoreUtils::standardizeFormula()]. Inputs can be
+#' either a `character` or `data.frame` containing a column with formulas.
+#' In case of `data.frame`s parameter `formulaColname` needs to be used to
+#' specify the name of the column containing the chemical formulas.
 #'
-#' @param query chemical formulas to search
+#' @param BPPARAM parallel processing setup. See `BiocParallel::bpparam()` for
+#'     details.
 #'
-#' @param target compound table to compare against
+#' @param formulaColname `character` with the name of the column containing
+#'     chemical formulas. Can be of length 1 if both `query` and `target` are
+#'     `data.frame`s and the name of the column with chemical formulas is the
+#'     same for both. If different columns are used, `formulaColname[1]` can be
+#'     used to define the column name in `query` and `formulaColname[2]` the
+#'     one of `target`.
 #'
-#' @param formulaColname name of the column contain chemical formula. Only used
-#'     if a `data.frame` is supplied.
+#' @param target `character` or `data.frame` with chemical formulas to compare
+#'     against.
+#'
+#' @param query `character` or `data.frame` with chemical formulas to search.
 #'
 #' @param ... currently ignored
 #'
-#' @return [Matched] object representing the result
+#' @return [Matched] object representing the result.
 #'
 #' @author Michael Witting
 #'
@@ -62,35 +71,34 @@ setGeneric("matchFormula", function(query, target, ...)
 #' @rdname matchFormula
 #'
 #' @importFrom MetaboCoreUtils standardizeFormula
-setMethod("matchFormula",
-          signature = c(query = "character",
-                        target = "character"),
-          function(query, target, BPPARAM = SerialParam()) {
-            query <- unname(standardizeFormula(query))
-            target <- unname(standardizeFormula(target))
-            target_formula <- data.frame(index = rep(seq_along(target)),
-                                         formula = target)
-            matches <- do.call(
-              rbind, bpmapply(seq_along(query), query, FUN = .getFormulaMatches,
-                              MoreArgs = list(target = target_formula),
-                              BPPARAM = BPPARAM, SIMPLIFY = FALSE))
-            Matched(query = query, target = target, matches = matches)
+setMethod(
+    "matchFormula",
+    signature = c(query = "character",
+                  target = "character"),
+    function(query, target, BPPARAM = SerialParam()) {
+        matches <- do.call(
+            rbind, bpmapply(seq_along(query), standardizeFormula(query),
+                            FUN = .getFormulaMatches,
+                            MoreArgs =list(target = standardizeFormula(target)),
+                            BPPARAM = BPPARAM, SIMPLIFY = FALSE))
+        Matched(query = query, target = target, matches = matches)
+    })
 
-          })
 #' @rdname matchFormula
 setMethod("matchFormula",
           signature = c(query = "data.frameOrSimilar",
                         target = "data.frameOrSimilar"),
           function(query, target, formulaColname = c("formula", "formula"),
                    BPPARAM = SerialParam()) {
-            if(length(formulaColname) == 1)
-              formulaColname <- rep(formulaColname, 2)
+            if(length(formulaColname) == 1L)
+              formulaColname <- rep(formulaColname, 2L)
             if(!formulaColname[1] %in% colnames(query))
-              stop("Missing column \"", formulaColname[1], "\" in query")
-            if(!formulaColname[2] %in% colnames(target))
-              stop("Missing column \"", formulaColname[2], "\" in target")
-            res <- matchFormula(query[, formulaColname[1]],
-                                target[, formulaColname[2]])
+              stop("Missing column \"", formulaColname[1L], "\" in query")
+            if(!formulaColname[2L] %in% colnames(target))
+              stop("Missing column \"", formulaColname[2L], "\" in target")
+            res <- matchFormula(query[, formulaColname[1L]],
+                                target[, formulaColname[2L]],
+                                BPPARAM = BPPARAM)
             res@query <- query
             res@target <- target
             res
@@ -101,10 +109,11 @@ setMethod("matchFormula",
           signature = c(query = "character",
                         target = "data.frameOrSimilar"),
           function(query, target, formulaColname = "formula",
-                   BBPARAM = SerialParam()) {
+                   BPPARAM = SerialParam()) {
             if(!formulaColname %in% colnames(target))
               stop("Missing column \"", formulaColname, "\" in target")
-            res <- matchFormula(query, target[, formulaColname])
+            res <- matchFormula(query, target[, formulaColname],
+                                BPPARAM = BPPARAM)
             res@query <- query
             res@target <- target
             res
@@ -115,10 +124,11 @@ setMethod("matchFormula",
           signature = c(query = "data.frameOrSimilar",
                         target = "character"),
           function(query, target, formulaColname = "formula",
-                   BBPARAM = SerialParam()) {
+                   BPPARAM = SerialParam()) {
             if(!formulaColname %in% colnames(query))
               stop("Missing column \"", formulaColname, "\" in query")
-            res <- matchFormula(query[, formulaColname], target)
+            res <- matchFormula(query[, formulaColname], target,
+                                BPPARAM = BPPARAM)
             res@query <- query
             res@target <- target
             res
@@ -131,21 +141,18 @@ setMethod("matchFormula",
 #'
 #' @param queryFormula `character(1)` with the formula of the query
 #'
-#' @param target `data.frame` with columns `"index"` and  `"formula"`
+#' @param target `character` with target formulas
 #'
 #' @noRd
 .getFormulaMatches <- function(queryIndex, queryFormula, target) {
-
-  cls <- which(queryFormula == target$formula)
-
-  if(length(cls)) {
-    data.frame(query_idx = queryIndex,
-               target_idx = cls,
-               score = 1)
-  } else {
-    data.frame(query_idx = integer(),
-              target_idx = integer(),
-              score = numeric())
-  }
+    cls <- which(target == queryFormula)
+    if(length(cls)) {
+        data.frame(query_idx = queryIndex,
+                   target_idx = cls,
+                   score = 1)
+    } else {
+        data.frame(query_idx = integer(),
+                   target_idx = integer(),
+                   score = numeric())
+    }
 }
-